requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
requests==2.26.0
moviepy==1.0.3
numpy==2.2.4
opencv-python-headless==4.11.0.86
Pillow==11.1.0
faster-whisper==1.1.1
werkzeug==2.0.1 